feat: add authorization logic

+++ practice/notes-app-v1/src/services/notes/controller/note-controller.js
@@ -2,8 +2,10 @@ import { nanoid } from 'nanoid';
 // import notes from '../notes.js';
 import InvariantError from '../../../exceptions/invariant-error.js';
 import NotFoundError from '../../../exceptions/not-found-error.js';
+import AuthorizationError from '../../../exceptions/authorization-error.js';
 import response from '../../../utils/response.js';
 import noteRepositories from '../repositories/note-repositories.js';
+// import AuthorizationError from '../../../exceptions/authorization-error.js';
 
 // export const addNoteHandler = (req, res, next) => {
 //   const { title = 'untitled', tags, body } = req.body;
@@ -84,7 +86,8 @@ import noteRepositories from '../repositories/note-repositories.js';
 
 export const addNoteHandler = async (req, res, next) => {
   const { title, body, tags } = req.body;
-  const note = await noteRepositories.createNote({ title, body, tags });
+  const { id: owner } = req.user;
+  const note = await noteRepositories.createNote({ title, body, tags, owner });
   if (!note) {
     return next(new InvariantError('Catatan gagal ditambahkan'));
   }
@@ -92,31 +95,55 @@ export const addNoteHandler = async (req, res, next) => {
 };
 
 export const getAllNotesHandler = async (req, res) => {
-  const noteDatas = await noteRepositories.getNotes();
+  const { id: owner } = req.user;
+  const noteDatas = await noteRepositories.getNotes(owner);
   return response(res, 200, 'Catatan sukses ditampilkan', { notes: noteDatas });
 };
 
 export const getNoteByIdHandler = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await noteRepositories.getNoteById(id);
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
+
   return response(res, 200, 'Catatan sukses ditampilkan', { note });
 };
 
 export const editNoteByIdHandler = async (req, res, next) => {
   const { id } = req.params;
-  const { title, body, tags } = req.body;
+  const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await noteRepositories.editNote({ id, title, body, tags });
   if (!note) {
     return next(new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan'));
   }
+
   return response(res, 200, 'Catatan berhasil diperbarui', { note });
 };
 
 export const deleteNoteByIdHandler = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const noteId = await noteRepositories.deleteNote(id);
   if (!noteId) {
     return next(new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan'));
